+++ R/yf_utils.R
@@ -193,3 +193,51 @@ check_internet <- function() {
 
   return(flag)
 }
+
+
+#' Convert bytes to a more natural representation
+#'
+#' Copied from https://github.com/gerrymanoim/humanize/blob/master/R/filesize.R
+#'
+#' @noRd
+natural_size <- function(bytes, suffix_type="decimal", fmt='%.1f') {
+
+  suffixes <- list(
+    'decimal' =  c('kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'),
+    'binary' = c('KiB', 'MiB', 'GiB', 'TiB', 'PiB', 'EiB', 'ZiB', 'YiB'),
+    'gnu' = c("K","M","G","T","P","E","Z","Y")
+  )
+
+  stopifnot(suffix_type %in% names(suffixes))
+  # How much value check do I need for bytes?
+  suffix <- suffixes[[suffix_type]]
+  gnu <- suffix_type == "gnu"
+
+  base <- ifelse(suffix_type %in% c('gnu', 'binary'), 1024, 1000)
+
+  if (bytes == 1 & !gnu) {
+    return("1 Byte")
+  } else if (bytes < base & !gnu) {
+    return(glue::glue("{bytes} Bytes"))
+  } else if (bytes < base & gnu) {
+    return(glue::glue("{bytes}B"))
+  }
+
+  for (i in seq_along(suffix)) {
+    unit <- base ^ (i + 1)
+    if (bytes < unit) {
+      out_val <- sprintf(fmt,(base * bytes / unit))
+      if (gnu) {
+        return(glue::glue("{out_val}{suffix[[i]]}"))
+      } else {
+        return(glue::glue("{out_val} {suffix[[i]]}"))
+      }
+    }
+  }
+
+  out_val <- sprintf(fmt,(base * bytes / unit))
+  if (gnu) {
+    return(glue::glue("{out_val}{suffix[[length(suffix)]]}"))
+  }
+  return(glue::glue("{out_val} {suffix[[length(suffix)]]}"))
+}